refactor: export db and auth from config
